Final Push - Ready for Review
File name changes and added author and date. A random picture will appear when the page loads. The picture can be changed upon clicking a figure element, which has styles to make it appear in the bottom left. Clicking the figure runs a stylesheet so that the image that was clicked appears.